# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Compliance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pat-track/config:
    get:
      operationId: getPATTrackConfig
      tags: [pat-track]
      summary: Get PAT testing configuration
      responses:
        "200":
          description: PAT testing configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PATTrackConfig"
    put:
      operationId: updatePATTrackConfig
      tags: [pat-track]
      summary: Update PAT testing configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PATTrackConfig"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PATTrackConfig"
  /pat-track/appliances:
    get:
      operationId: listPATAppliances
      tags: [pat-track]
      summary: List PAT appliances with their latest test summary
      responses:
        "200":
          description: List of appliances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PATAppliance"
    post:
      operationId: createPATAppliance
      tags: [pat-track]
      summary: Add a PAT appliance to the register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePATApplianceRequest"
      responses:
        "201":
          description: Created appliance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PATAppliance"
  /pat-track/appliances/{id}:
    put:
      operationId: updatePATAppliance
      tags: [pat-track]
      summary: Update a PAT appliance
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePATApplianceRequest"
      responses:
        "200":
          description: Updated appliance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PATAppliance"
    delete:
      operationId: deletePATAppliance
      tags: [pat-track]
      summary: Delete a PAT appliance
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /pat-track/tests:
    get:
      operationId: listPATTests
      tags: [pat-track]
      summary: List PAT test records
      parameters:
        - in: query
          name: applianceId
          schema:
            type: integer
      responses:
        "200":
          description: List of test records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PATTest"
    post:
      operationId: createPATTest
      tags: [pat-track]
      summary: Record a PAT test result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePATTestRequest"
      responses:
        "201":
          description: Created test record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PATTest"
  /pat-track/tests/{id}:
    put:
      operationId: updatePATTest
      tags: [pat-track]
      summary: Update a PAT test record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePATTestRequest"
      responses:
        "200":
          description: Updated test record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PATTest"
    delete:
      operationId: deletePATTest
      tags: [pat-track]
      summary: Delete a PAT test record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /pat-track/preset-templates/{key}:
    put:
      operationId: savePATPresetTemplate
      tags: [pat-track]
      summary: Save a customised appliance preset template
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/PATPresetItem"
              required: [items]
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    delete:
      operationId: deletePATPresetTemplate
      tags: [pat-track]
      summary: Reset a preset template back to built-in defaults
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    PATAppliance:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        siteId:
          type: integer
          nullable: true
        name:
          type: string
        applianceType:
          type: string
        location:
          type: string
          nullable: true
        assetTag:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        active:
          type: boolean
        lastTestDate:
          type: string
          nullable: true
        lastResult:
          type: string
          nullable: true
        nextTestDate:
          type: string
          nullable: true
        lastTestedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - name
        - applianceType
        - active
    CreatePATApplianceRequest:
      type: object
      properties:
        name:
          type: string
        applianceType:
          type: string
        location:
          type: string
          nullable: true
        assetTag:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        siteId:
          type: integer
          nullable: true
        active:
          type: boolean
      required:
        - name
    PATTest:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        applianceId:
          type: integer
        testDate:
          type: string
        result:
          type: string
          enum: [pass, fail]
        nextTestDate:
          type: string
          nullable: true
        testedBy:
          type: string
          nullable: true
        visualInspection:
          type: string
          nullable: true
          enum: [pass, fail, na]
        earthContinuityOhms:
          type: string
          nullable: true
        insulationMohms:
          type: string
          nullable: true
        operatingCurrent:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        applianceName:
          type: string
          nullable: true
        assetTag:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - applianceId
        - testDate
        - result
    CreatePATTestRequest:
      type: object
      properties:
        applianceId:
          type: integer
        testDate:
          type: string
        result:
          type: string
          enum: [pass, fail]
        nextTestDate:
          type: string
          nullable: true
        testedBy:
          type: string
          nullable: true
        visualInspection:
          type: string
          nullable: true
          enum: [pass, fail, na]
        earthContinuityOhms:
          type: string
          nullable: true
        insulationMohms:
          type: string
          nullable: true
        operatingCurrent:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - applianceId
        - testDate
        - result
...
    PATPresetItem:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
      required:
        - name
        - type
...
    PATTrackConfig:
      type: object
      properties:
        pat_show_earth_bond:
          type: string
        pat_show_insulation:
          type: string
        pat_default_tester:
          type: string
        pat_retest_months:
          type: string
        pat_locations:
          type: string
...